chore(Loader component): rename loading to loader add Loader to carousel page
